refactor: move day schedule to separate component
